+++ src/components/pages/_BasePage.jsx
@@ -4,7 +4,7 @@ import Header from '../layouts/Header';
 import Footer from '../layouts/Footer';
 import StudentProfile from '../../components/pages/StudentProfile';
 import Joblistingjobs from '../pages/Joblistingjobs'
-import FormApply from '../core-ui/formAppy'
+import FormApply from '../core-ui/FormAppy'
 import EditStudentAccount from '../pages/EditStudentAccount'
 import Account from '../pages/Account'
 export default function _BasePage() {
