+++ src/styles/components/__index.scss
@@ -17,5 +17,5 @@
 @import "./footer";
 @import "./pnvPartners.scss";
 @import "account";
-@import "__forms";
+@import "forms";
 @import "editaccount";
